pyproject.toml
[project]
name = "twitter-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "jupyterlab",
    "ipykernel",
    "ipywidgets",
    "openai-agents",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "tweepy>=4.15.0",
]
